Flow Edit - Transformations tab : check if Abs function works as expected
Create new transformation function (Absolute function) then run the transformation to see the result of this function and check the values to verify  if Abs works as expected and thats done by using " Math.abs( ) ".
Take the value for the First Elemnt in the  row and calculate the Abs to this value then compare this value to The value generated by the function itself in the site.